+++ app.js
@@ -1,8 +1,12 @@
 let Axios = require('axios');
 let WebSocketAsPromised = require('websocket-as-promised');
 let W3CWebSocket = require('websocket').w3cwebsocket;
-let logger = require('./logger');
 let striptags = require('striptags');
+let Sentry = require('@sentry/node');
+let { SENTRY_DSN_KEY } = require('./local');
+
+// Initialize Sentry client
+Sentry.init({ dsn: SENTRY_DSN_KEY });
 
 // Config for websocket-as-promised
 const wsp = new WebSocketAsPromised('ws://localhost:6700', {
@@ -74,6 +78,7 @@ async function scanOfficialAccountPost(user_id, user_name) {
         lastScanTimestamps[user_id] = thisScanTimestamp;
     } catch(err) {
         logger.log({ level: 'error', message: err });
+        Sentry.captureException(err);
     }
 }
 
@@ -91,6 +96,7 @@ async function scanMemberPost(user_id, user_name) {
         lastScanTimestamps[user_id] = thisScanTimestamp;
     } catch(err) {
         logger.log({ level: 'error', message: err });
+        Sentry.captureException(err);
     }
 }
 
@@ -124,6 +130,7 @@ async function sendWebsocketMessage(message) {
         wsp.send(JSON.stringify(makePayload(GroupChatIds.TaskForce)));
     } catch(err) {
         logger.log({ level: 'error', message: err });
+        Sentry.captureException(err);
     } finally {
         await wsp.close();
     }
@@ -140,4 +147,4 @@ async function main() {
 
 setInterval(function() {
     main();
-}, 20000);
+}, 60000);